requirements.txt
discord.py==2.3.2
google-generativeai==0.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
flask==2.3.3
aiohttp==3.8.5